aws_lambda_functions/fetchTeeTimesPY/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.12.12
tenacity==8.2.3
pytz==2024.1
supabase==2.3.4 